Moving Unit Test to Integration Test. Adding more tests. Removing unnecessary `ANALYTICS_EVENTS_REQUIRE_AUTHENTICATION` env var.
